后端基本完成
后端基本已经完成了，只剩下控制器部分，前端计划用jsp+elemtUI

+++ src/main/java/entity/Coach.java
@@ -8,5 +8,5 @@ public class Coach {
   private String userId;
   private String workId;
   private String carSchoolId;
-  private long appraiseCoach;
+  private int appraiseCoach;
 }
